requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.3.0
requests>=2.31.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
marshmallow>=3.20.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
